Conflicts:
	clig/accelsearch.1
	clig/prepfold.1
	docs/accelsearch.1
	docs/prepfold.1
	include/accel.h
	include/presto.h
	src/accel_utils.c
	src/accelsearch_cmd.c
	src/maximize_rz.c
	src/prepfold_cmd.c

+++ src/barycenter.c
@@ -15,10 +15,11 @@ int read_resid_rec(FILE * file, double *toa, double *obsf)
 /* the file resid2.tmp which is written by TEMPO.       */
 /* It returns 1 if successful, 0 if unsuccessful.       */
 {
-   static int l;
+   static long l;
    static double d[9];
 
-   fread(&l, sizeof(int), 1, file);
+   // For g77, the following should be sizeof(long)
+   fread(&l, sizeof(long), 1, file);
    fread(&d, sizeof(double), 9, file);
    //printf("Barycentric TOA = %17.10f\n", d[0]);
    //printf("Postfit residual (pulse phase) = %g\n", d[1]);
@@ -31,7 +32,8 @@ int read_resid_rec(FILE * file, double *toa, double *obsf)
    //printf("??? = %g\n\n", d[8]);
    *toa = d[0];
    *obsf = d[4];
-   return fread(&l, sizeof(int), 1, file);
+   // For g77, the following should be sizeof(long)
+   return fread(&l, sizeof(long), 1, file);
 }
 
 void barycenter(double *topotimes, double *barytimes,
@@ -109,9 +111,9 @@ void barycenter(double *topotimes, double *barytimes,
    }
    fclose(outfile);
 
-   /* rename("itoa.out", "itoa1.out"); */
-   /* rename("bary.tmp", "bary1.tmp"); */
-   /* rename("bary.par", "bary1.par"); */
+   //rename("itoa.out", "itoa1.out");
+   //rename("bary.tmp", "bary1.tmp");
+   //rename("bary.par", "bary1.par");
 
    /* Write the free format TEMPO file to begin barycentering */
 
@@ -169,9 +171,9 @@ void barycenter(double *topotimes, double *barytimes,
 
    /* Cleanup the temp files */
 
-   /* rename("itoa.out", "itoa2.out"); */
-   /* rename("bary.tmp", "bary2.tmp"); */
-   /* rename("bary.par", "bary2.par"); */
+   //rename("itoa.out", "itoa2.out");
+   //rename("bary.tmp", "bary2.tmp");
+   //rename("bary.par", "bary2.par");
 
    remove("tempo.lis");
    remove("tempoout_times.tmp");
